Refactor Checkpoint/Restore Infrastructure
Add a new class, TR::CRRuntime, to encapsulate all the of the
checkpoint/restore code.
